+++ front-api/routes/poke/workspaces/[wId]/projects/index.ts
@@ -1,17 +1,9 @@
-import {
-  listAllProjectsWithAdminMetadata,
-  type ProjectWithAdminMetadata,
-} from "@app/lib/api/projects/list";
+import type { PokeListProjects } from "@app/lib/api/poke/projects";
+import { listAllProjectsWithAdminMetadata } from "@app/lib/api/projects/list";
 import { pokeApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 
 import projectId from "./[projectId]";
-
-export type PokeProjectType = ProjectWithAdminMetadata;
-
-export type PokeListProjects = {
-  projects: PokeProjectType[];
-};
 
 // Mounted at /api/poke/workspaces/:wId/projects.
 const app = pokeApp();
