+++ front-api/routes/poke/workspaces/[wId]/triggers/[tId]/webhook_requests.ts
@@ -1,23 +1,11 @@
+import type { PokeGetWebhookRequestsResponseBody } from "@app/lib/api/poke/triggers";
 import { TriggerResource } from "@app/lib/resources/trigger_resource";
 import { fetchRecentWebhookRequestTriggersWithPayload } from "@app/lib/triggers/webhook";
-import type { WebhookRequestTriggerStatus } from "@app/types/assistant/triggers";
 import { WEBHOOK_REQUEST_TRIGGER_STATUSES } from "@app/types/assistant/triggers";
 import { pokeApp } from "@front-api/middlewares/ctx";
 import { apiError, type HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { z } from "zod";
-
-export interface PokeGetWebhookRequestsResponseBody {
-  requests: {
-    id: number;
-    timestamp: number;
-    status: WebhookRequestTriggerStatus;
-    payload?: {
-      headers?: Record<string, string | string[]>;
-      body?: unknown;
-    };
-  }[];
-}
 
 const WebhookRequestsQuerySchema = z.object({
   limit: z.coerce.number().int().positive().optional(),
