+++ front/pages/api/poke/workspaces/[wId]/conversations/index.ts
@@ -1,25 +1,17 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
+import type {
+  PokeListConversationItem,
+  PokeListConversations,
+} from "@app/lib/api/poke/conversations";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { apiError } from "@app/logger/withlogging";
-import type {
-  ConversationVisibility,
-  ConversationWithoutContentType,
-} from "@app/types/assistant/conversation";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import { isString } from "@app/types/shared/utils/general";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeListConversationItem = ConversationWithoutContentType & {
-  visibility?: ConversationVisibility;
-};
-
-export type PokeListConversations = {
-  conversations: PokeListConversationItem[];
-};
 
 async function handler(
   req: NextApiRequest,
