+++ front/pages/api/poke/workspaces/[wId]/triggers/[tId]/execution_stats.ts
@@ -1,6 +1,7 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
+import type { PokeGetTriggerExecutionStats } from "@app/lib/api/poke/triggers";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { TriggerResource } from "@app/lib/resources/trigger_resource";
@@ -9,17 +10,6 @@ import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import { isString } from "@app/types/shared/utils/general";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeGetTriggerExecutionStats = {
-  statusBreakdown: Record<string, number>;
-  dailyVolume: Array<{
-    date: string;
-    succeeded: number;
-    failed: number;
-    notMatched: number;
-    rateLimited: number;
-  }>;
-};
 
 async function handler(
   req: NextApiRequest,
