+++ front/poke/swr/conversation.ts
@@ -1,5 +1,5 @@
+import type { PokeListConversations } from "@app/lib/api/poke/conversations";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeListConversations } from "@app/pages/api/poke/workspaces/[wId]/conversations";
 import type { PokeConditionalFetchProps } from "@app/poke/swr/types";
 import type { Fetcher } from "swr";
