+++ front/poke/swr/conversation_config.ts
@@ -1,5 +1,5 @@
+import type { PokeGetConversationConfig } from "@app/lib/api/poke/conversations";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeGetConversationConfig } from "@app/pages/api/poke/workspaces/[wId]/conversations/[cId]/config";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
