+++ front/poke/swr/project_tasks.ts
@@ -1,5 +1,5 @@
+import type { PokeListProjectTasks } from "@app/lib/api/poke/projects";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeListProjectTasks } from "@app/pages/api/poke/workspaces/[wId]/projects/[projectId]/tasks";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
