+++ front/poke/swr/projects.ts
@@ -1,5 +1,5 @@
+import type { PokeListProjects } from "@app/lib/api/poke/projects";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeListProjects } from "@app/pages/api/poke/workspaces/[wId]/projects";
 import type { PokeConditionalFetchProps } from "@app/poke/swr/types";
 import type { Fetcher } from "swr";
