+++ front/poke/swr/trigger_execution_stats.ts
@@ -1,5 +1,5 @@
+import type { PokeGetTriggerExecutionStats } from "@app/lib/api/poke/triggers";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeGetTriggerExecutionStats } from "@app/pages/api/poke/workspaces/[wId]/triggers/[tId]/execution_stats";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
